Make final TS adjustments
