+++ check/index.d.ts
@@ -1,4 +1,4 @@
-export { Validator, ValidationChain } from './check';
+export { Validator, ValidationChain, ValidatorOptions } from './check';
 export * from './one-of';
 export * from './validation-chain-builders';
 export * from './validation-result';
