+++ index.d.ts
@@ -5,14 +5,13 @@
 
 ///<reference types="express"/>
 import * as express from 'express';
+import { Location } from './check/location';
 import {
-  Dictionary,
   Result,
-  MappedError,
-  Options,
-  Sanitizer,
-  Validator as BaseValidator
-} from './shared-typings';
+  Validator as BaseValidator,
+  ValidatorOptions
+} from './check';
+import { Sanitizer } from './filter'
 
 // Add RequestValidation Interface on to Express's Request Interface.
 declare global {
@@ -20,46 +19,104 @@ declare global {
     interface Request extends ExpressValidator.RequestValidation { }
   }
 }
+
 export as namespace ExpressValidator;
-/**
- * @param options see: https://github.com/ctavan/express-validator#middleware-options
- * @constructor
- */
-declare function ExpressValidator(options?: Options.ExpressValidatorOptions): express.RequestHandler;
+interface ExpressValidatorOptions {
+  errorFormatter?(param: string, msg: string, value: any, location: Location): any,
+  customSanitizers?: Record<string, (...value: any[]) => any>,
+  customValidators?: Record<string, (...value: any[]) => boolean | Promise<any>>,
+}
+
+declare function ExpressValidator(options?: ExpressValidatorOptions): express.RequestHandler;
 export = ExpressValidator;
+
 // Internal Module.
 declare namespace ExpressValidator {
+  export type ValidationSchema = Record<
+    string,
+    ValidationSchemaParamOptions | { [custom: string]: ValidatorSchemaOptions }
+  >;
 
-  export type URLProtocol = 'http' | 'https' | 'ftp'
-  export type UUIDVersion = 3 | 4 | 5 | 'all'
-  export type IPVersion = 4 | 6
-  export type AlphaLocale = 'ar' | 'ar-AE' | 'ar-BH' | 'ar-DZ' | 'ar-EG' | 'ar-IQ' | 'ar-JO' | 'ar-KW' | 'ar-LB' | 'ar-LY' | 'ar-MA' | 'ar-QA' | 'ar-QM' | 'ar-SA' | 'ar-SD' | 'ar-SY' | 'ar-TN' | 'ar-YE' | 'cs-CZ' | 'da-DK' | 'de-DE' | 'en-AU' | 'en-GB' | 'en-HK' | 'en-IN' | 'en-NZ' | 'en-US' | 'en-ZA' | 'en-ZM' | 'es-ES' | 'fr-FR' | 'hu-HU' | 'nl-NL' | 'pl-PL' | 'pt-BR' | 'pt-PT' | 'ru-RU' | 'sr-RS' | 'sr-RS@latin' | 'tr-TR' | 'uk-UA'
-  export type AlphanumericLocale = 'ar' | 'ar-AE' | 'ar-BH' | 'ar-DZ' | 'ar-EG' | 'ar-IQ' | 'ar-JO' | 'ar-KW' | 'ar-LB' | 'ar-LY' | 'ar-MA' | 'ar-QA' | 'ar-QM' | 'ar-SA' | 'ar-SD' | 'ar-SY' | 'ar-TN' | 'ar-YE' | 'cs-CZ' | 'da-DK' | 'de-DE' | 'en-AU' | 'en-GB' | 'en-HK' | 'en-IN' | 'en-NZ' | 'en-US' | 'en-ZA' | 'en-ZM' | 'es-ES' | 'fr-FR' | 'fr-BE' | 'hu-HU' | 'nl-BE' | 'nl-NL' | 'pl-PL' | 'pt-BR' | 'pt-PT' | 'ru-RU' | 'sr-RS' | 'sr-RS@latin' | 'tr-TR' | 'uk-UA'
-  export type MobilePhoneLocal = 'ar-DZ' | 'ar-SA' | 'ar-SY' | 'cs-CZ' | 'de-DE' | 'da-DK' | 'el-GR' | 'en-AU' | 'en-GB' | 'en-HK' | 'en-IN' | 'en-NZ' | 'en-US' | 'en-CA' | 'en-ZA' | 'en-ZM' | 'es-ES' | 'fi-FI' | 'fr-FR' | 'hu-HU' | 'it-IT' | 'ja-JP' | 'ms-MY' | 'nb-NO' | 'nn-NO' | 'pl-PL' | 'pt-PT' | 'ru-RU' | 'sr-RS' | 'tr-TR' | 'vi-VN' | 'zh-CN' | 'zh-TW'
-  export type Location = 'body' | 'params' | 'query' | 'headers' // TODO add cookies if #292 is accepted
+  interface ValidatorSchemaOptions {
+    options?: any[]
+    errorMessage?: any
+  }
 
-  export type ValidationSchema = {
-    [param: string]:
-    Options.ValidationSchemaParamOptions // standard validators
-    | // or
-    { [customValidator: string]: Options.ValidatorSchemaOptions } // custom ones
+  interface ValidationSchemaParamOptions {
+    in?: Location
+    errorMessage?: any
+
+    // Additional ValidatorChain.prototype.* validators
+    optional?: boolean | ValidatorOptions.OptionalOptions
+    notEmpty?: boolean | { errorMessage: string }
+    len?: ValidatorSchemaOptions
+
+    // exported from validator.js
+    isEmail?: ValidatorSchemaOptions
+    isURL?: ValidatorSchemaOptions
+    isMACAddress?: ValidatorSchemaOptions
+    isIP?: ValidatorSchemaOptions
+    isFQDN?: ValidatorSchemaOptions
+    isBoolean?: ValidatorSchemaOptions
+    isAlpha?: ValidatorSchemaOptions
+    isAlphanumeric?: ValidatorSchemaOptions
+    isNumeric?: ValidatorSchemaOptions
+    isLowercase?: ValidatorSchemaOptions
+    isUppercase?: ValidatorSchemaOptions
+    isAscii?: ValidatorSchemaOptions
+    isFullWidth?: ValidatorSchemaOptions
+    isHalfWidth?: ValidatorSchemaOptions
+    isVariableWidth?: ValidatorSchemaOptions
+    isHash?: ValidationSchemaParamOptions
+    isMultibyte?: ValidatorSchemaOptions
+    isSurrogatePair?: ValidatorSchemaOptions
+    isInt?: ValidatorSchemaOptions
+    isFloat?: ValidatorSchemaOptions
+    isDecimal?: ValidatorSchemaOptions
+    isHexadecimal?: ValidatorSchemaOptions
+    isDivisibleBy?: ValidatorSchemaOptions
+    isHexColor?: ValidatorSchemaOptions
+    isMD5?: ValidatorSchemaOptions
+    isJSON?: ValidatorSchemaOptions
+    isLatLong?: ValidationSchemaParamOptions
+    isEmpty?: ValidatorSchemaOptions
+    isLength?: ValidatorSchemaOptions
+    isByteLength?: ValidatorSchemaOptions
+    isPostalCode?: ValidationSchemaParamOptions
+    isUUID?: ValidatorSchemaOptions
+    isMongoId?: ValidatorSchemaOptions
+    isDate?: ValidatorSchemaOptions
+    isAfter?: ValidatorSchemaOptions
+    isBefore?: ValidatorSchemaOptions
+    isIn?: ValidatorSchemaOptions
+    isCreditCard?: ValidatorSchemaOptions
+    isISIN?: ValidatorSchemaOptions
+    isISRC?: ValidatorSchemaOptions
+    isISBN?: ValidatorSchemaOptions
+    isISSN?: ValidatorSchemaOptions
+    isMobilePhone?: ValidatorSchemaOptions
+    isCurrency?: ValidatorSchemaOptions
+    isISO8601?: ValidatorSchemaOptions
+    isBase64?: ValidatorSchemaOptions
+    isDataURI?: ValidatorSchemaOptions
+    isWhitelisted?: ValidatorSchemaOptions
+
+    // Additional Validators provided by validator.js
+    equals?: ValidatorSchemaOptions
+    contains?: ValidatorSchemaOptions
+    matches?: ValidatorSchemaOptions
   }
 
   interface ValidatorFunction {
-    (item: string | string[], message?: any): Validator;
+    (item: string | string[] | number, message?: any): Validator;
     (schema: {}): Validator;
   }
-  /**
-   * This one's used for RegexRoutes
-   * @see https://github.com/ctavan/express-validator#regex-routes
-   */
-  interface ValidatorFunctionRegExp extends ValidatorFunction { (matchIndex: number, message?: string): Validator; }
   interface SanitizerFunction { (item: string): Sanitizer; }
 
   export interface RequestValidation {
-    assert: ValidatorFunctionRegExp;
-    validate: ValidatorFunctionRegExp;
-    check: ValidatorFunctionRegExp;
+    assert: ValidatorFunction;
+    validate: ValidatorFunction;
+    check: ValidatorFunction;
     checkBody: ValidatorFunction;
     checkCookies: ValidatorFunction;
     checkHeaders: ValidatorFunction;
@@ -74,39 +131,16 @@ declare namespace ExpressValidator {
     sanitizeHeaders: SanitizerFunction;
     sanitizeCookies: SanitizerFunction;
 
-    /**
-     * @deprecated
-     * @param mapped Will cause the validator to return an object that maps parameter to error.
-     * @return Synchronous errors in the form of an array. If there are no errors, the returned value is false.
-     */
-    validationErrors(mapped?: boolean): Dictionary<MappedError> | MappedError[];
-    /**
-     * @deprecated
-     * @param mapped Will cause the validator to return an object that maps parameter to error.
-     * @return Synchronous errors in the form of an array. If there are no errors, the returned value is false.
-     */
-    validationErrors<T>(mapped?: boolean): Dictionary<T> | T[];
-    /**
-     * @deprecated
-     * @param mapped Whether to map parameters to errors or not.
-     */
-    asyncValidationErrors(mapped?: boolean): Promise<MappedError[] | Dictionary<MappedError>>;
-    /**
-     * @deprecated
-     * @param mapped Whether to map parameters to errors or not.
-     */
-    asyncValidationErrors<T>(mapped?: boolean): Promise<T[] | Dictionary<T>>;
-    /**
-     * @return Promise<Result> A Promise which resolves to a result object.
-     */
+    validationErrors(mapped?: boolean): Record<string, any> | any[];
+    validationErrors<T>(mapped?: boolean): Record<string, T> | T[];
+    asyncValidationErrors(mapped?: boolean): Promise<any[] | Record<string, any>>;
+    asyncValidationErrors<T>(mapped?: boolean): Promise<T[] | Record<string, T>>;
     getValidationResult(): Promise<Result>
   }
 
-  export interface Validator extends BaseValidator {
-    // Additional legacy validators
-
+  export interface Validator extends BaseValidator, Sanitizer {
     notEmpty(): this;
-    len(options: Options.MinMaxOptions): this;
+    len(options: ValidatorOptions.MinMaxOptions): this;
   }
 
 }
