+++ tsconfig.json
@@ -9,7 +9,6 @@
     "noImplicitAny": true,
     "noImplicitReturns": true,
     "noImplicitThis": true,
-    "noUnusedLocals": true,
     "noUnusedParameters": true,
     "strictNullChecks": true,
     "outDir": "./build"
